pyproject.toml
[tool.poetry]
name = "adboard-api"
version = "0.1.0"
description = ""
authors = ["vitek <rgavih@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.5.0"
drf-yasg = "^1.21.8"
load-dotenv = "^0.1.0"
black = "^24.10.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = "migrations"